+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+- Update file size for images after processing
 
 - Display avatar correctly cropped in message metadata view
 - End URLs on whitespace